Compute index for method; method exists from index
Rather than compute method exists directly, the library now computes
the index of the method then determines whether the method exists
based on the index.  This does slow down the general case, but 1)
there are more improvements to be made; 2) it now allows for an
even faster scenario where the method index is cached for repeated
use; 3) it will allow safe proxy to use method exists more
quickly and easily.