Fix thread issue with GetMethodIndex
Since the proxy created by ComputeMethodIndexMixin uses state
to compute the method index, it was subject to returning the
wrong value if used from more than one thread.  MethodIndexFactory
now combines this with a thread static StaticInstanceMixin so
that each thread gets its own instance.  Access via the property
was slow, so a static access func was added to StaticInstanceMixin