Add package projects based on ASP.NET 5 RC1
Code will now compile under ASP.NET 5 RC1 Update 1 using the right
set of precompiler flags and will run on the dotnet5.4 framework.
The new package projects can be used to make a nuget package with
.NET 4.0, 4.5, and dotnet5.4.  .NET 4.0 csproj files remain to
allow for use with the Resharper test runner and use with
peverify (to be revisted with dotnet cli).